Allow external plugins  - borrows code brom ElectronCash  - external plugins are imported as zip files  - plugins are enabled per wallet  - 'run_hook' is now a wallet method
Note that there is no guarantee of insulation between wallets;
if a plugin is enabled by a wallet, its code may affect other
wallets that have been loaded by the user.

Hardware wallets and trustedcoin modules are still defined as
plugins. It would make sense to move them out of the plugin
system, because they need to be loaded in the wizard. This
is not done in this PR; for the moment, these plugins use
'run_wizard_hook', which does not require a wallet instance.